# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.TerminalRequest:
    properties:
      address:
        type: string
      id:
        type: integer
      name:
        type: string
      serial_number:
        type: string
    type: object
  handlers.TransactionRequest:
    properties:
      amount:
        type: number
      card_id:
        type: integer
      id:
        type: integer
      terminal_id:
        type: integer
    type: object
...
  repository.Terminal:
    properties:
      address:
        type: string
      id:
        type: integer
      name:
        type: string
      serialNumber:
        type: string
    type: object
  repository.Transaction:
    properties:
      amount:
        type: number
      cardID:
        type: integer
      createdAt:
        type: string
      id:
        type: integer
      terminalID:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  title: Lab2 REST API
  version: "1.0"
paths:
...
  /terminals/all:
    get:
      description: Returns all payment terminals.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/repository.Terminal'
            type: array
        "405":
          description: Method not allowed
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: List terminals
      tags:
      - terminals
  /terminals/create:
    post:
      consumes:
      - application/json
      description: Creates a terminal. Admin JWT required (is_admin=1).
      parameters:
      - description: Bearer admin JWT
        in: header
        name: Authorization
        required: true
        type: string
      - description: Terminal fields
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.TerminalRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Validation error
          schema:
            type: string
        "401":
          description: Missing or invalid token
          schema:
            type: string
        "403":
          description: Not an admin
          schema:
            type: string
        "405":
          description: Method not allowed
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create terminal
      tags:
      - terminals
  /terminals/delete:
    delete:
      description: Deletes a terminal by id. Admin JWT required.
      parameters:
      - description: Bearer admin JWT
        in: header
        name: Authorization
        required: true
        type: string
      - description: Terminal ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid id
          schema:
            type: string
        "401":
          description: Missing or invalid token
          schema:
            type: string
        "403":
          description: Not an admin
          schema:
            type: string
        "405":
          description: Method not allowed
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete terminal
      tags:
      - terminals
  /terminals/get:
    get:
      description: Returns a single terminal by numeric id.
      parameters:
      - description: Terminal ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repository.Terminal'
        "400":
          description: Invalid id or not found
          schema:
            type: string
        "405":
          description: Method not allowed
          schema:
            type: string
      summary: Get terminal by ID
      tags:
      - terminals
  /terminals/update:
    put:
      consumes:
      - application/json
      description: Updates a terminal by id. Admin JWT required.
      parameters:
      - description: Bearer admin JWT
        in: header
        name: Authorization
        required: true
        type: string
      - description: Must include id and fields
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.TerminalRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Validation error
          schema:
            type: string
        "401":
          description: Missing or invalid token
          schema:
            type: string
        "403":
          description: Not an admin
          schema:
            type: string
        "405":
          description: Method not allowed
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update terminal
      tags:
      - terminals
  /transactions/all:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/repository.Transaction'
            type: array
        "405":
          description: Method not allowed
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: List transactions
      tags:
      - transactions
  /transactions/create:
    post:
      consumes:
      - application/json
      description: Admin only. Records amount for a card at a terminal.
      parameters:
      - description: Bearer admin JWT
        in: header
        name: Authorization
        required: true
        type: string
      - description: amount, card_id, terminal_id
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.TransactionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Validation error
          schema:
            type: string
        "401":
          description: Missing or invalid token
          schema:
            type: string
        "403":
          description: Not an admin
          schema:
            type: string
        "405":
          description: Method not allowed
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create transaction
      tags:
      - transactions
  /transactions/delete:
    delete:
      description: Admin only.
      parameters:
      - description: Bearer admin JWT
        in: header
        name: Authorization
        required: true
        type: string
      - description: Transaction ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid id
          schema:
            type: string
        "401":
          description: Missing or invalid token
          schema:
            type: string
        "403":
          description: Not an admin
          schema:
            type: string
        "405":
          description: Method not allowed
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete transaction
      tags:
      - transactions
  /transactions/get:
    get:
      parameters:
      - description: Transaction ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repository.Transaction'
        "400":
          description: Invalid id or not found
          schema:
            type: string
        "405":
          description: Method not allowed
          schema:
            type: string
      summary: Get transaction by ID
      tags:
      - transactions
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"